Extract `runtime` module
