+++ srml/contract/src/vm/prepare.rs
@@ -231,7 +231,7 @@ mod tests {
 	fn parse_and_prepare_wat(wat: &str) -> Result<PreparedContract, Error> {
 		let wasm = wabt::Wat2Wasm::new().validate(false).convert(wat).unwrap();
 		let config = Config::<Test>::default();
-		let env = ::vm::env_def::init_env();
+		let env = ::vm::runtime::init_env();
 		prepare_contract::<MockExt>(wasm.as_ref(), &config, &env)
 	}
